pyproject.toml
[tool.poetry]
name = "secure-agentic-soc-copilot"
version = "0.1.0"
description = "A secure, agentic assistant for Security Operations Centers"
authors = ["Antigravity <antigravity@deepmind.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
langchain = "^0.1.0"
langgraph = "^0.0.10"
chromadb = "^0.4.22"
redis = "^5.0.1"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-exporter-otlp = "^1.22.0"
gpt4all = "^2.1.0"
python-multipart = "^0.0.6"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
